+++ packager/pom.xml
@@ -18,12 +18,12 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 <modelVersion>4.0.0</modelVersion>
 
-  <groupId>co.cask.marketplace</groupId>
-  <artifactId>cask-packager</artifactId>
+  <groupId>io.cdap.hub</groupId>
+  <artifactId>hub-packager</artifactId>
   <version>0.1.0-SNAPSHOT</version>
-  <name>Cask Marketplace Packager</name>
-  <description>Tools to create and publish packages to a CDAP Marketplace.</description>
-  <url>https://github.com/caskdata/cask-marketplace</url>
+  <name>CDAP Hub Packager</name>
+  <description>Tools to create and publish packages to a CDAP Hub.</description>
+  <url>https://github.com/cdapio/hub</url>
 
   <licenses>
     <license>
@@ -36,17 +36,17 @@
 
   <developers>
     <developer>
-      <name>Cask Data</name>
-      <email>cask-dev@googlegroups.com</email>
-      <organization>Cask Data, Inc.</organization>
+      <name>CDAP</name>
+      <email>cdap-dev@googlegroups.com</email>
+      <organization>CDAP</organization>
       <organizationUrl>http://www.cdap.io</organizationUrl>
     </developer>
   </developers>
 
   <scm>
-    <connection>scm:git:https://github.com/caskdata/cask-marketplace.git</connection>
-    <developerConnection>scm:git:git@github.com:caskdata/cask-marketplace.git</developerConnection>
-    <url>https://github.com/caskdata/cask-marketplace.git</url>
+    <connection>scm:git:https://github.com/cdapio/hub.git</connection>
+    <developerConnection>scm:git:git@github.com:cdapio/hub.git</developerConnection>
+    <url>https://github.com/cdapio/hub.git</url>
     <tag>HEAD</tag>
   </scm>
