+++ packager/src/main/java/io/cdap/hub/spec/ActionSpec.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.marketplace.spec;
+package io.cdap.hub.spec;
 
 import java.util.List;
 import java.util.Objects;
